refactored ingredients and fetch call
